헤드, 스토어, 창고 모두 서버 연결 (기존 스토어와 창고 생성자 삭제)
